代码生成以及规格列表

+++ pinyougou-dao/src/main/java/com/pinyougou/mapper/TbAreasMapper.java
@@ -2,15 +2,16 @@ package com.pinyougou.mapper;
 
 import com.pinyougou.pojo.TbAreas;
 import com.pinyougou.pojo.TbAreasExample;
-import java.util.List;
 import org.apache.ibatis.annotations.Param;
+
+import java.util.List;
 
 public interface TbAreasMapper {
     int countByExample(TbAreasExample example);
 
     int deleteByExample(TbAreasExample example);
 
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(TbAreas record);
 
@@ -18,7 +19,7 @@ public interface TbAreasMapper {
 
     List<TbAreas> selectByExample(TbAreasExample example);
 
-    TbAreas selectByPrimaryKey(Integer id);
+    TbAreas selectByPrimaryKey(Long id);
 
     int updateByExampleSelective(@Param("record") TbAreas record, @Param("example") TbAreasExample example);
