+++ pinyougou-dao/src/main/java/com/pinyougou/mapper/TbProvincesMapper.java
@@ -2,15 +2,16 @@ package com.pinyougou.mapper;
 
 import com.pinyougou.pojo.TbProvinces;
 import com.pinyougou.pojo.TbProvincesExample;
-import java.util.List;
 import org.apache.ibatis.annotations.Param;
+
+import java.util.List;
 
 public interface TbProvincesMapper {
     int countByExample(TbProvincesExample example);
 
     int deleteByExample(TbProvincesExample example);
 
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(TbProvinces record);
 
@@ -18,7 +19,7 @@ public interface TbProvincesMapper {
 
     List<TbProvinces> selectByExample(TbProvincesExample example);
 
-    TbProvinces selectByPrimaryKey(Integer id);
+    TbProvinces selectByPrimaryKey(Long id);
 
     int updateByExampleSelective(@Param("record") TbProvinces record, @Param("example") TbProvincesExample example);
